Upgrade to LLVM 20.1.0
I just published an LLVM 20.1.0 toolchain with some computed gotos
patches to hopefully improve performance of CPython.

https://github.com/indygreg/toolchain-tools/releases/tag/toolchain-bootstrap%2F20250308

One of the patches changes codegen to fix a regression in LLVM 19.

The other enables bolt to work with computed gotos. To take full
advantage of this functionality we patch CPython configure to remove
some `--skip-funcs` which are no longer required.