Replace various *proto packages with xorgproto
As part of trying to stand up native Linux ARM builds, I ran into a
failure with xproto's config.guess not recognizing ARM. That package
not releasing since 2016 clued me in that it was likely not a required
part of X11 any more.

I looked at Debian's dependency tree and there was no xproto in sight.
But there was an xorgproto. And sure enough if you install xorgproto
instead of xproto in the build environment, things "just work."

Clued into the potential for outdated packages, I looked at other
proto packages. I pulled the thread and inputproto, kbproto, and
xextproto all appeared to be dead and replaced by xorgproto.

So this commit deletes inputproto, kbproto, xextproto, and xproto
and effectively replaces their use with xorgproto.

Good riddance.