Update quirks doc to match current reality
We configure ncurses with defaults for TERMINFO_DIRS that work on I can
find (it works on Debian family, Red Hat family, NixOS, and Alpine), so
we shouldn't phrase either the title or the body as if this is a problem
that affects most people or setting the environment variable is a
necessary thing in general. (Also I think even in cases where we don't
have terminfo for your terminal, backspace usually still works, as
evidenced by `TERM=potato python`).

musl distributions are now dynamic binaries by default and dynamic
loading works.

The libcrypt transition is behind us.

The python3-config file (since its first version, I think) knows how to
correct the prefix in output. While it does contain hard-coded paths
internally, it uses that for the purpose of fixing its output, so I
don't think that fact is helpful to call out.

python-build-standalone users who get it through uv are (hopefully) not
affected by absolute paths in sysconfig; call this out explicitly for
the benefit of people evaluting whether uv will work for them. Also
mention sysconfigpatcher for everyone else.

In general, tighten up the content.