Disable unsafe identical code folding in BOLT
astral-sh/uv#13610 reported a misbehavior that is the result of a subclass of
str incorrectly having its ->tp_as_number->nb_add slot filled in with the value
of PyUnicode_Type->tp_as_sequence->sq_concat. There are some times when this is
an appropriate thing to do iwhen subclassing, but this is not one of them. The
logic to prevent it in this case relies on two helper functions in the file,
wrap_binaryfunc and wrap_binaryfunc_l, having different addresses, even though
they contain identical code.

For some reason BOLT does not do this optimization in the shared library
(even though those are static functions and not exported), so we only
started seeing this in the static build.