Add zlib as a build-dep of Tcl/Tk
This allows the shared libraries to statically link libz.a. I think we
were previously picking up the zlib headers from the OS, and we were
previously building static libraries so that was okay - they got linked
into CPython itself which had the zlib symbols. But now that we're
building shared libraries, they're picking up a dependency on libz.so.1,
presumably because we don't have a static zlib installed from the OS.