Give _tkinter $ORIGIN-relative dependencies on glibc and an rpath on musl
Partially addresses #742 and makes it consistent with what we're doing
on macOS.

There's an argument in the comments above that we should not set an
rpath on libpython (except on musl where it's needed), and I need to see
if I still believe that. In the meantime I'm following that pattern and
setting $ORIGIN-relative NEEDED on glibc and rpath on musl only.

This also adds a specific ldd regression test but not the additional
tests listed in #742.