sqlite: Remove -DSQLITE_ENABLE_FTS3_TOKENIZER and add tests for compile options
As noted in the discussion in #562, compiling SQLite with the
-DSQLITE_ENABLE_FTS3_TOKENIZER flag is equivalent to using
`connection.setconfig(sqlite3.SQLITE_DBCONFIG_ENABLE_FTS3_TOKENIZER)` at
runtime. The purpose of this option, in either syntax, is to disable a
security measure to provide backwards compatibility for older code.
Specifically, the `fts3_tokenizer()` function can accept or return a
native-code pointer to a structure containing callback functions, which
makes it an attractive target for SQL injection attacks to escalate to
arbitrary native code execution. The more-secure behavior is to require
the use of bound parameters with this function; the backwards-compatible
behavior allows the function to be called with blob literals or computed
values. Because of a documentation shortcoming, some applications
thought they needed this option on at compile time, and so Debian's
SQLite build, used by e.g. the `python` container on Dockerhub, has it
on. But there is no functionality that is only enabled by having this
option on at compile time. Ideally, applications should use bound
parameters when calling this function. If that code change is hard, they
can alternatively set the option themselves at runtime to preserve
compatibility with existing code, but that still doesn't need anything
turned on at compile time. So the right decision for us is not to enable
this flag at compile time and preserve the secure behavior.

Add a test that `fts3_tokenizer()` is usable with bound parameters but
not with blob literals, and also add tests for a couple of other
preivously-requested SQLite flags for compatibility with other
implementations:

* #309: -DSQLITE_ENABLE_DBSTAT_VTAB
* #449: serialize/deserialize (on by default, was just a compile-time detection issue)
* #550: -DSQLITE_ENABLE_FTS3_PARENTHESIS