macOS CI: Stop setting $APPLE_SDK_PATH
The build script supports locating this with `xcrun --show-sdk-path`. It
looks like we originally started overriding this in CI to force use of
the 10.15 SDK to work around issues with the 11.0 one (09ca2ac775). We
should be able to support building with the newest SDK because we set
`-mmacosx-version-min` / `MACOSX_DEPLOYMENT_TARGET` to a low value, and
this keeps us robust to GitHub runner image changes. There are SDK
compatibility tests, too.