Fix musl detection in Python 3.12 and below
CPython 3.10 and below don't even attempt to fix the target triple for
musl. CPython 3.11 and 3.12 do, but in our setup it doens't work right
because of an autoconf bug. CPython 3.13 rewrites the target triple
detection logic and works fine.

Meanwhile, the previous patch to this part of configure.ac is no longer
relevant because clang stopped implementing --print-multiarch in LLVM 16
(llvm/llvm-project@e05edb19), so we can drop those patches.

Fixes #724.